refactor: move coordinator module into coordinator/ package (Phase 7 of #625)
