+++ custom_components/localshift/computation_engine_lib/excess_solar_signals.py
@@ -16,7 +16,7 @@ from ..const import (
     DEFAULT_DEMAND_WINDOW_START,
     DEFAULT_MINIMUM_TARGET_SOC,
 )
-from ..coordinator_data import CoordinatorData
+from ..coordinator.data import CoordinatorData
 
 _LOGGER = logging.getLogger(__name__)
