+++ custom_components/localshift/coordinator/__init__.py
@@ -1,0 +1,15 @@
+"""Coordinator package for LocalShift integration."""
+
+from .coordinator import LocalShiftCoordinator
+from .data import (
+    AdaptiveParameters,
+    CoordinatorData,
+    PerformanceMetrics,
+)
+
+__all__ = [
+    "AdaptiveParameters",
+    "CoordinatorData",
+    "LocalShiftCoordinator",
+    "PerformanceMetrics",
+]
