+++ custom_components/localshift/integration/controller.py
@@ -19,7 +19,7 @@ from .client import PowerwallServiceClient
 if TYPE_CHECKING:
     from homeassistant.core import HomeAssistant
 
-    from ..coordinator_data import CoordinatorData
+    from ..coordinator.data import CoordinatorData
 
 
 _LOGGER = logging.getLogger(__name__)
