+++ tests/computation_engine_lib/test_forecast_history_store.py
@@ -9,7 +9,7 @@ from homeassistant.util import dt as dt_util
 from custom_components.localshift.computation_engine_lib.forecast_history_store import (
     ForecastHistoryStore,
 )
-from custom_components.localshift.coordinator_data import CoordinatorData
+from custom_components.localshift.coordinator import CoordinatorData
 
 
 @pytest.fixture
