+++ tests/learning/test_learning_integration.py
@@ -23,7 +23,7 @@ from custom_components.localshift.computation_engine_lib.pattern_analyzer import
     PatternAnalyzer,
 )
 from custom_components.localshift.const import BatteryMode
-from custom_components.localshift.coordinator_data import (
+from custom_components.localshift.coordinator import (
     AdaptiveParameters,
     CoordinatorData,
     PerformanceMetrics,
