+++ tests/test_optimizer_facade.py
@@ -5,7 +5,7 @@ from datetime import UTC, datetime
 from custom_components.localshift.computation_engine_lib.optimizer_facade import (
     OptimizerFacade,
 )
-from custom_components.localshift.coordinator_data import CoordinatorData
+from custom_components.localshift.coordinator import CoordinatorData
 
 
 class _StubSlotBuilder:
